% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{mc_check_comp}
\alias{mc_check_comp}
\alias{mc_fit_comp}
\alias{mc_models_comp}
\alias{mc_predict_comp}
\alias{subset_CPI_data}
\alias{pred_density_moose_CPI}
\title{Check Composition Data}
\usage{
mc_check_comp(x)

mc_fit_comp(x, vars = NULL)

mc_models_comp(model_list_comp)

mc_predict_comp(
  total_model_id,
  comp_model_id,
  model_list_total,
  model_list_comp,
  x,
  do_avg = FALSE
)

subset_CPI_data(CPI, ss)

pred_density_moose_CPI(CPI, ...)
}
\arguments{
\item{x}{data frame}

\item{vars}{column names of `x` to be used as predictors for the composition model}

\item{model_list_comp}{named list of total composition models}

\item{total_model_id}{model ID or model IDs for total moose model (can be multiple model IDs from `names(model_list_total)`)}

\item{comp_model_id}{model ID or model IDs for composition model (single model ID from `names(model_list_comp)`)}

\item{model_list_total}{named list of total moose models}

\item{do_avg}{logical, to do model averaging}

\item{CPI}{Composition PI output}

\item{ss}{a subset of rows (logical or numeric vector)}

\item{...}{other arts passed to print}
}
\description{
Check the composition data frame to make sure composition columns add up and give total Moose as a result.

Multinomial model to capture how predictors affect compsotional data.

Extract useful summaries from a mlist of composition models.

Calculate prediction intervals based on composition model.
}
