% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{explore}
\alias{explore}
\alias{mc_plot_univariate}
\alias{mc_plot_multivariate}
\alias{mc_plot_comp}
\title{Exploration}
\usage{
mc_plot_univariate(
  i,
  x,
  dist = "ZINB",
  base = TRUE,
  type = c("density", "map", "fit"),
  interactive = FALSE
)

mc_plot_multivariate(vars, x, alpha = NULL)

mc_plot_comp(i, x)
}
\arguments{
\item{i}{Column name from `x` to be used as a predictor.}

\item{x}{Data frame with Moose data.}

\item{dist}{Count distribution (`P`, `NB`, `ZIP`, or `ZINB`).}

\item{base}{Logical, draw base graphics or ggplot2.}

\item{type}{Character, type of plot to be drawn (`"density"`, `"map"`, `"fit"`). Base plot can draw all 3, ggplot2 can only draw one at a time.}

\item{interactive}{Logical, draw interactive plot (not available for base plots).}

\item{vars}{A vector of column names from `x` to be used as a predictor.}

\item{alpha}{Alpha level defining `mincriterion = 1 - alpha` for `partykit::ctree()`.}
}
\description{
Functions to explore relationships between total Moose or composition
as respone vs. environmental predictor variables.
}
\details{
`mc_plot_univariate` implements visual univariate
(sigle predictor) exploration for the total Moose count models.

`mc_plot_multivariate` implements visual multivariate
(multiple predictors) exploration based on regression trees
(recursive partitioning  in a conditional inference framework)
for total Moose counts.

`mc_plot_comp` implements visual univariate (sigle predictor) exploration
for the multinomial composition models.
}
\examples{
## Prepare Moose data from Mayo
x <- read.csv(
    system.file("extdata/MayoMMU_QuerriedData.csv",
        package="moosecounter"))
switch_response("total")
x <- mc_update_total(x)

## Univariate exploration for total Moose
mc_plot_univariate("Subalp_Shrub_250buf", x, "ZINB")

## Multivariate exploration for total Moose
vars <- c("ELC_Subalpine", "Fire1982_2012", "Fire8212_DEM815",
    "NALC_Needle", "NALC_Shrub", "Subalp_Shrub_250buf",
    "ELCSub_Fire8212DEM815", "SubShrub250_Fire8212DEM815")
mc_plot_multivariate(vars, x)

## Univariate exploration for composition
mc_plot_comp("Fire8212_DEM815", x)

}
\keyword{models}
\keyword{regression}
\keyword{tree}
