% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/models.R
\name{rZINB}
\alias{rZINB}
\alias{solvenear}
\alias{find_mode}
\alias{\%>\%}
\alias{wzi}
\alias{zeroinfl2}
\title{Internal functions}
\usage{
rZINB(N, mu.nb, theta.nb, phi.zi)

solvenear(x)

find_mode(x)

lhs \%>\% rhs

wzi(object, pass_data = FALSE, ...)

zeroinfl2(
  formula,
  data,
  subset,
  na.action,
  weights,
  offset,
  dist = c("ZIP", "ZINB", "P", "NB"),
  link = c("logit", "probit", "cloglog"),
  control = NULL,
  model = TRUE,
  y = TRUE,
  x = FALSE,
  solveH = TRUE,
  ...
)
}
\arguments{
\item{N}{the number of random numbers to produce}

\item{mu.nb}{the mean of the NB distribution}

\item{theta.nb}{the dispersion parameter of the NB distribution, a single non-negative numeric value or `NULL` (this refers to no overdispersion, thus a Poisson or ZIP distribution)}

\item{phi.zi}{the probability of the non-zero valies in the ZI part}

\item{x}{a numeric vector}

\item{object}{a model as returned by `zeroinfl2()`}

\item{pass_data}{logical, to pass the data or not}

\item{...}{control arguments}

\item{formula}{model formula as in `pscl::zeroinfl()`}

\item{data}{Moose data set}

\item{subset, na.action, weights, offset, model, y, x}{arguments as in `pscl::zeroinfl()`}

\item{dist}{count distribution, one of `"ZIP"`, `"ZINB"`, `"P"`, `"NB"`}

\item{link}{link function for the zero model}

\item{control}{see `pscl::zeroinfl.control()`}

\item{solveH}{logical, to use robust matrix inversion to get VCV}

\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
A numeric vector of length `N` with random numbers

The result of calling `rhs(lhs)`.
}
\description{
`ZINB` makes zero inflated negative binomial random numbers.

`solvenear` inverts near-PD matrices.
It function makes sure that near-positive definite matrices
are positive definite. Positive definiteness is needed
for matrix inversion, which in turn is used to find
the Hessian matrix and standard errors for model coefficients
from numerical optimization.

The `find_mode` function finds the mode of a distribution using
one-dimensional kernel density estimation.
The density based estimate is rounded, because
the function is used in the context of count data models
and predictions.

`\%>\%`: imported function,
see \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.

`wzi` applies a leave-one-out approach to temper influential observations.
The process finds weights that are related to leverage
(how much each observation contributes to the model likelihood).

`zeroinfl2` is a customized version of the `pscl::zeroinfl()` function, but this also fits the non-ZI counterparts in a way that simplifies downstream analyses (i.e. PI calculations). Intended for internal use.
}
\examples{
table(rZINB(100, 3, 2, 0.5))
table(rZINB(100, 3, NULL, 0.5))
x <- c(1, 2, 1, 3, 4, 3, 2, 3, 5, 6, 10)
find_mode(x)
plot(density(x))
rug(x)
abline(v = find_mode(x), lty=2)

}
\keyword{internal}
