% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{mc_check_comp}
\alias{mc_check_comp}
\title{Check Composition Data}
\usage{
mc_check_comp(x)
}
\arguments{
\item{x}{Moose data object}
}
\description{
Check the composition data frame to make sure composition columns add up and give total Moose as a result.
}
