% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{mc_fit_comp}
\alias{mc_fit_comp}
\title{Fit Composition Model for Total Moose}
\usage{
mc_fit_comp(x, vars = NULL)
}
\arguments{
\item{x}{data frame}

\item{vars}{column names of `x` to be used as predictors for the composition model}
}
\description{
Fit Composition Model for Total Moose
}
