% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{mc_fit_total}
\alias{mc_fit_total}
\title{Fit Model to Total Abundance}
\usage{
mc_fit_total(vars, x, zi_vars = NULL, dist = "ZINB", weighted = FALSE, ...)
}
\arguments{
\item{vars}{column names of `x` to be used as predictors for the count model}

\item{x}{data frame (use `mc_update_total()` on `x` prior)}

\item{zi_vars}{optional, column names of `x` to be used as predictors for the zero model}

\item{dist}{distribution (P, NB, ZIP, ZINB)}

\item{weighted}{logical, to use weighting to moderate influential observations}

\item{...}{other args passed to `zeroinfl2()`}
}
\description{
Fit Model to Total Abundance
}
