% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gassaway.R
\name{mc_gassaway}
\alias{mc_gassaway}
\alias{print.mc_gassaway}
\title{Total Estimate with Gassaway Formula}
\usage{
mc_gassaway(y1, y2, N1, N2)

\method{print}{mc_gassaway}(x, ...)
}
\arguments{
\item{y1, y2}{surveyed cell-level population counts in the 2 strata}

\item{N1, N2}{total number of cells that were stratified}

\item{x}{object to print}

\item{...}{arguments passed to the print method}
}
\description{
Total Estimate with Gassaway Formula
}
\examples{
y1 <- rpois(20, 50)
y2 <- rpois(30, 5)
N1 <- 40
N2 <- 50
mc_gassaway(y1, y2, N1, N2)

}
