% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{mc_get_pred}
\alias{mc_get_pred}
\title{Get PI Info}
\usage{
mc_get_pred(PI, ss = NULL)
}
\arguments{
\item{PI}{PI object returned by `mc_predict_total()`}

\item{ss}{subset indicator}
}
\description{
Get PI Info
}
