% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{mc_models_comp}
\alias{mc_models_comp}
\title{Composition Model Table}
\usage{
mc_models_comp(model_list_comp)
}
\arguments{
\item{model_list_comp}{composition model list}
}
\description{
Composition Model Table
}
