% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{mc_models_total}
\alias{mc_models_total}
\title{Total Abundance Models}
\usage{
mc_models_total(ml, x, coefs = TRUE)
}
\arguments{
\item{ml}{named list of models}

\item{x}{data frame}

\item{coefs}{logical, return coefficient table too}
}
\description{
Total Abundance Models
}
