% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{mc_plot_comp}
\alias{mc_plot_comp}
\title{Plot Univariate Composition Model}
\usage{
mc_plot_comp(i, x)
}
\arguments{
\item{i}{predictor variable name in `x`}

\item{x}{moose data frame}
}
\description{
Plot Univariate Composition Model
}
