% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mc_plot_multivariate}
\alias{mc_plot_multivariate}
\title{Multivariate Exploration}
\usage{
mc_plot_multivariate(vars, x)
}
\arguments{
\item{vars}{column names from `x` to be used as a predictor}

\item{x}{data frame with Moose data}
}
\description{
Multivariate Exploration
}
