% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{mc_plot_pidistr}
\alias{mc_plot_pidistr}
\title{Plot the total PI distribution}
\usage{
mc_plot_pidistr(PI, plot = TRUE, breaks = "Sturges")
}
\arguments{
\item{PI}{PI object returned by `mc_predict_total()`}

\item{plot}{logical, to plot or just give summary}

\item{breaks}{breaks arg passed to `graphics::hist()`}
}
\description{
Plot the total PI distribution
}
