% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{mc_plot_pidistrcell}
\alias{mc_plot_pidistrcell}
\title{Plot the total PI distribution for a cell}
\usage{
mc_plot_pidistrcell(PI, id = 1, plot = TRUE, breaks = "Sturges")
}
\arguments{
\item{PI}{PI object returned by `mc_predict_total()`}

\item{id}{cell ID}

\item{plot}{logical, to plot or just give summary}

\item{breaks}{breaks arg passed to `graphics::hist()`}
}
\description{
Plot the total PI distribution for a cell
}
