% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{mc_plot_residuals}
\alias{mc_plot_residuals}
\title{Plot Model Residuals}
\usage{
mc_plot_residuals(model_id, ml, x)
}
\arguments{
\item{model_id}{model ID (one from `names(ml)`)}

\item{ml}{named list of models}

\item{x}{data frame}
}
\description{
Plot Model Residuals
}
