% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mc_plot_univariate}
\alias{mc_plot_univariate}
\title{Univariate Exploration}
\usage{
mc_plot_univariate(i, x, dist = "ZINB")
}
\arguments{
\item{i}{column name from `x` to be used as a predictor}

\item{x}{data frame with Moose data}

\item{dist}{count distribution (P, NB, ZIP, or ZINB)}
}
\description{
Univariate exploration for the total Moose models.
}
