% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{mc_predict_comp}
\alias{mc_predict_comp}
\title{Composition Prediction Intervals}
\usage{
mc_predict_comp(
  total_model_id,
  comp_model_id,
  model_list_total,
  model_list_comp,
  x,
  do_avg = FALSE
)
}
\arguments{
\item{total_model_id}{model ID or model IDs for total moose model (can be multiple model IDs from `names(model_list_total)`)}

\item{comp_model_id}{model ID or model IDs for composition model (single model ID from `names(model_list_comp)`)}

\item{model_list_total}{named list of total moose models}

\item{model_list_comp}{named list of total composition models}

\item{x}{data frame}

\item{do_avg}{logical, to do model averaging}
}
\description{
Calculate prediction intervals based on composition model.
}
