% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{mc_predict_total}
\alias{mc_predict_total}
\title{Total Abundance Prediction Intervals}
\usage{
mc_predict_total(model_id, ml, x, do_boot = TRUE, do_avg = FALSE)
}
\arguments{
\item{model_id}{model ID or model IDs (can be multiple from `names(ml)`)}

\item{ml}{named list of models}

\item{x}{data frame}

\item{do_boot}{logical, to do bootstrap}

\item{do_avg}{logical, to do model averaging}
}
\description{
Total Abundance Prediction Intervals
}
