% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{mc_update_total}
\alias{mc_update_total}
\title{Update Moose Data}
\usage{
mc_update_total(x, srv = NULL, ss = NULL)
}
\arguments{
\item{x}{data frame}

\item{srv}{logical vector, rows of `x` that are surveyed, falls back to global options when `NULL`}

\item{ss}{logical vector to subset `x`, default is to take no subset}
}
\description{
Update Moose Data
}
