% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{wzi}
\alias{wzi}
\alias{zeroinfl2}
\alias{summary.non_zeroinfl}
\alias{print.summary.non_zeroinfl}
\alias{nobs.zeroinfl}
\title{Internal Functions: Models}
\usage{
wzi(object, pass_data = FALSE, ...)

zeroinfl2(
  formula,
  data,
  subset,
  na.action,
  weights,
  offset,
  dist = c("ZIP", "ZINB", "P", "NB"),
  link = c("logit", "probit", "cloglog"),
  control = NULL,
  model = TRUE,
  y = TRUE,
  x = FALSE,
  solveH = TRUE,
  ...
)

\method{summary}{non_zeroinfl}(object, ...)

\method{print}{summary.non_zeroinfl}(x, digits = max(3, getOption("digits") - 3), ...)

\method{nobs}{zeroinfl}(object, ...)
}
\arguments{
\item{object}{a model as returned by `zeroinfl2()`}

\item{pass_data}{logical, to pass the data or not}

\item{...}{control arguments}

\item{formula}{model formula as in `pscl::zeroinfl()`}

\item{data}{Moose data set}

\item{subset, na.action, weights, offset, model, y, x}{arguments as in `pscl::zeroinfl()`}

\item{dist}{count distribution, one of `"ZIP"`, `"ZINB"`, `"P"`, `"NB"`}

\item{link}{link function for the zero model}

\item{control}{see `pscl::zeroinfl.control()`}

\item{x}{A model as returned by `zeroinfl2()`.}

\item{solveH}{logical, to use robust matrix inversion to get VCV}

\item{digits}{Digits for print method.}
}
\description{
`wzi` applies a leave-one-out approach to temper influential observations.
The process finds weights that are related to leverage
(how much each observation contributes to the model likelihood).

`zeroinfl2` is a customized version of the `pscl::zeroinfl()` function, but this also fits the non-ZI counterparts in a way that simplifies downstream analyses (i.e. PI calculations). Intended for internal use.
}
\keyword{internal}
\keyword{models}
\keyword{regression}
