% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{pred_density_moose_CPI}
\alias{pred_density_moose_CPI}
\title{Print Composition PI Results}
\usage{
pred_density_moose_CPI(CPI, ...)
}
\arguments{
\item{CPI}{Composition PI output}

\item{...}{other arts passed to print}
}
\description{
Print Composition PI Results
}
