% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{pred_density_moose_PI}
\alias{pred_density_moose_PI}
\title{Get PI Density}
\usage{
pred_density_moose_PI(PI)
}
\arguments{
\item{PI}{PI object returned by `mc_predict_total()`}
}
\description{
Get PI Density
}
