% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rZINB}
\alias{rZINB}
\title{Random Numbers From ZINB Distribution}
\usage{
rZINB(N, mu.nb, theta.nb, phi.zi)
}
\arguments{
\item{N}{the number of random numbers to produce}

\item{mu.nb}{the mean of the NB distribution}

\item{theta.nb}{the dispersion parameter of the NB distribution, a single non-negative numeric value or `NULL` (this refers to no overdispersion, thus a Poisson or ZIP distribution)}

\item{phi.zi}{the probability of the non-zero valies in the ZI part}
}
\value{
A numeric vector of length `N` with random numbers
}
\description{
ZINB = zero inflated negative binomial.
}
\examples{
table(rZINB(100, 3, 2, 0.5))
table(rZINB(100, 3, NULL, 0.5))
}
