% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{solvenear}
\alias{solvenear}
\title{Robust Matrix Inversion}
\usage{
solvenear(x)
}
\arguments{
\item{x}{a symmetric square matrix}
}
\description{
This function makes sure that near-positive definite matrices
are positive definite. Positive definiteness is needed
for matrix inversion, which in turn is used to find
the Hessian matrix and standard errors for model coefficients
from numerical optimization.
}
