% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{subset_CPI_data}
\alias{subset_CPI_data}
\title{Subset composition PI data object}
\usage{
subset_CPI_data(CPI, ss)
}
\arguments{
\item{CPI}{Composition PI output}

\item{ss}{a subset of rows (logical or numeric vector)}
}
\description{
Subset composition PI data object
}
