% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{switch_response}
\alias{switch_response}
\title{Switch Response}
\usage{
switch_response(type = "total")
}
\arguments{
\item{type}{type of the response, can be `"total"` or `"cows"`}
}
\description{
Switch between total Moose vs. cows only.
This sets the column name for totel Moose estimation.
}
\examples{

switch_response("cows")

}
