% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{switch_response}
\alias{switch_response}
\alias{mc_update_total}
\alias{mc_fit_total}
\alias{mc_models_total}
\alias{mc_predict_total}
\alias{mc_get_pred}
\alias{pred_density_moose_PI}
\alias{mc_plot_residuals}
\alias{mc_plot_predpi}
\alias{mc_plot_pidistr}
\title{Total Moose Workflow}
\usage{
switch_response(type = "total")

mc_update_total(x, srv = NULL, ss = NULL)

mc_fit_total(
  x,
  vars = NULL,
  zi_vars = NULL,
  dist = "ZINB",
  weighted = FALSE,
  ...
)

mc_models_total(ml, x, coefs = TRUE)

mc_predict_total(model_id, ml, x, do_boot = TRUE, do_avg = FALSE)

mc_get_pred(PI, ss = NULL)

pred_density_moose_PI(PI)

mc_plot_residuals(model_id, ml, x)

mc_plot_predpi(PI)

mc_plot_pidistr(PI, id = NULL, plot = TRUE, breaks = "Sturges")
}
\arguments{
\item{type}{The type of the response, can be `"total"` or `"cows"` for `switch_response`.}

\item{x}{data frame}

\item{srv}{Logical vector, rows of `x` that are surveyed, falls back to global options when `NULL`.}

\item{ss}{subset indicator}

\item{vars}{column names of `x` to be used as predictors for the count model}

\item{zi_vars}{optional, column names of `x` to be used as predictors for the zero model}

\item{dist}{distribution (P, NB, ZIP, ZINB)}

\item{weighted}{logical, to use weighting to moderate influential observations}

\item{...}{other args passed to `zeroinfl2()`}

\item{ml}{named list of models}

\item{coefs}{logical, return coefficient table too}

\item{model_id}{model ID (one from `names(ml)`)}

\item{do_boot}{logical, to do bootstrap}

\item{do_avg}{logical, to do model averaging}

\item{PI}{PI object returned by `mc_predict_total()`}

\item{id}{cell ID}

\item{plot}{logical, to plot or just give summary}

\item{breaks}{breaks arg passed to `graphics::hist()`}
}
\description{
`switch_response` switches between total Moose vs. cows only.
This sets the column name for totel Moose estimation.

`mc_update_total` Updates/prepares the Moose data set for downstream analyses (i.e. calculates some derived variables, sets a surveyed/unsurveyed indicator, and optinally takes a subset).

`mc_fit_total` fit total Moose abundance models.

`mc_models_total` prints out estimates from the models.
}
\examples{

switch_response("cows")

}
\keyword{models}
\keyword{regression}
