% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{wzi}
\alias{wzi}
\title{Weighted Refitting of Models}
\usage{
wzi(object, pass_data = FALSE, ...)
}
\arguments{
\item{object}{a model as returned by `zeroinfl2()`}

\item{pass_data}{logical, to pass the data or not}

\item{...}{other params}
}
\description{
Weighted Refitting of Models
}
