% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{zeroinfl2}
\alias{zeroinfl2}
\title{Count Models}
\usage{
zeroinfl2(
  formula,
  data,
  subset,
  na.action,
  weights,
  offset,
  dist = c("ZIP", "ZINB", "P", "NB"),
  link = c("logit", "probit", "cloglog"),
  control = NULL,
  model = TRUE,
  y = TRUE,
  x = FALSE,
  solveH = TRUE,
  ...
)
}
\arguments{
\item{formula}{model formula as in `pscl::zeroinfl()`}

\item{data}{Moose data set}

\item{subset, na.action, weights, offset, model, y, x}{arguments as in `pscl::zeroinfl()`}

\item{dist}{count distribution, one of `"ZIP"`, `"ZINB"`, `"P"`, `"NB"`}

\item{link}{link function for the zero model}

\item{control}{see `pscl::zeroinfl.control()`}

\item{solveH}{logical, to use robust matrix inversion to get VCV}

\item{...}{control arguments}
}
\description{
Count Models
}
