% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{comp}
\alias{comp}
\alias{mc_check_comp}
\alias{mc_fit_comp}
\alias{mc_models_comp}
\alias{mc_predict_comp}
\alias{subset_CPI_data}
\alias{pred_density_moose_CPI}
\title{Moose Composition Workflow}
\usage{
mc_check_comp(x)

mc_fit_comp(x, vars = NULL)

mc_models_comp(model_list_comp, coefs = TRUE)

mc_predict_comp(
  total_model_id,
  comp_model_id,
  model_list_total,
  model_list_comp,
  x,
  do_avg = FALSE
)

subset_CPI_data(CPI, ss)

pred_density_moose_CPI(CPI, ...)
}
\arguments{
\item{x}{A Moose data frame object.}

\item{vars}{Column names of `x` to be used as predictors
for the composition model.}

\item{model_list_comp}{Named list of total composition models.}

\item{coefs}{logical, return coefficient table too.}

\item{total_model_id}{Model ID or model IDs for total moose
model (can be multiple model IDs from `names(model_list_total)`).}

\item{comp_model_id}{Model ID or model IDs for composition model
(single model ID from `names(model_list_comp)`).}

\item{model_list_total}{Named list of total moose models.}

\item{do_avg}{Logical, to do model averaging or not.}

\item{CPI}{Composition PI object.}

\item{ss}{A subset of rows (logical or numeric vector).}

\item{...}{Other arts passed to underlying functions.}
}
\description{
Fit composition model for Moose using a multinomial model
to capture how predictors affect compsotional data;
alculate prediction intervals based on composition model;
and xtract useful summaries.
}
\examples{
mc_options(B=10)
x <- read.csv(
    system.file("extdata/MayoMMU_QuerriedData.csv",
        package="moosecounter"))
## Prepare Moose data frame object
x <- mc_update_total(x)

## Total moose model list
vars <- c("ELC_Subalpine", "Fire1982_2012", "Fire8212_DEM815",
    "NALC_Needle", "NALC_Shrub", "Subalp_Shrub_250buf",
    "ELCSub_Fire8212DEM815", "SubShrub250_Fire8212DEM815")
ML <- list()
ML[["Model 0"]] <- mc_fit_total(x, dist="ZINB", weighted=TRUE)
ML[["Model 1"]] <- mc_fit_total(x, vars[1:2], dist="ZINB", weighted=TRUE)
ML[["Model 2"]] <- mc_fit_total(x, vars[2:3], dist="ZIP", weighted=TRUE)
ML[["Model 3"]] <- mc_fit_total(x, vars[3:4], dist="ZINB", weighted=TRUE)

## Composition odel list
CML <- list()
CML[['FireDEMSub']] <- mc_fit_comp(x, "Fire8212_DEM815")

## Stats from the models
mc_models_comp(CML)

## Calculate PI
CPI <- mc_predict_comp(
    total_model_id="Model 3",
    comp_model_id='FireDEMSub',
    model_list_total=ML,
    model_list_comp=CML,
    x=x,
    do_avg=FALSE)

## Predict density
pred_density_moose_CPI(CPI)

}
\keyword{models}
\keyword{regression}
