% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{mc_plot_univariate}
\alias{mc_plot_univariate}
\alias{mc_plot_multivariate}
\alias{mc_plot_comp}
\title{Exploration}
\usage{
mc_plot_univariate(i, x, dist = "ZINB")

mc_plot_multivariate(vars, x, alpha = NULL)

mc_plot_comp(i, x)
}
\arguments{
\item{i}{predictor variable name in `x`}

\item{x}{moose data frame}

\item{dist}{count distribution (P, NB, ZIP, or ZINB)}

\item{vars}{column names from `x` to be used as a predictor}

\item{alpha}{alpha level defining mincriterion = 1 - alpha for `partykit::ctree()`}
}
\description{
`mc_plot_univariate` implements visual univariate
(sigle predictor) exploration for the total Moose count models.

`mc_plot_multivariate` implements visual multivariate
(multiple predictors) exploration based on regression trees
(recursive partitioning  in a conditional inference framework)
for total Moose counts.

`mc_plot_comp` implements visual univariate (sigle predictor) exploration
for the multinomial composition models.
}
\keyword{models}
\keyword{regression}
\keyword{tree}
