% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{rZINB}
\alias{rZINB}
\alias{solvenear}
\alias{find_mode}
\alias{\%>\%}
\title{Internal Functions: Helpers}
\usage{
rZINB(N, mu.nb, theta.nb, phi.zi)

solvenear(x)

find_mode(x)

lhs \%>\% rhs
}
\arguments{
\item{N}{the number of random numbers to produce}

\item{mu.nb}{the mean of the NB distribution}

\item{theta.nb}{the dispersion parameter of the NB distribution, a single non-negative numeric value or `NULL` (this refers to no overdispersion, thus a Poisson or ZIP distribution)}

\item{phi.zi}{the probability of the non-zero valies in the ZI part}

\item{x}{a symmetric square matrix for `solvenear`,
a numeric vector for `find_mode`}

\item{lhs}{A value or the magrittr placeholder.}

\item{rhs}{A function call using the magrittr semantics.}
}
\value{
A numeric vector of length `N` with random numbers

The result of calling `rhs(lhs)`.
}
\description{
`ZINB` makes zero inflated negative binomial random numbers.

`solvenear` inverts near-PD matrices.
It function makes sure that near-positive definite matrices
are positive definite. Positive definiteness is needed
for matrix inversion, which in turn is used to find
the Hessian matrix and standard errors for model coefficients
from numerical optimization.

The `find_mode` function finds the mode of a distribution using
one-dimensional kernel density estimation.
The density based estimate is rounded, because
the function is used in the context of count data models
and predictions.

`\%>\%`: imported function,
see \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\examples{
table(rZINB(100, 3, 2, 0.5))
table(rZINB(100, 3, NULL, 0.5))
x <- c(1, 2, 1, 3, 4, 3, 2, 3, 5, 6, 10)
find_mode(x)
plot(density(x))
rug(x)
abline(v = find_mode(x), lty=2)

}
\keyword{internal}
