% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp.R
\name{mc_fit_comp}
\alias{mc_fit_comp}
\title{Fit Composition Model for Total Moose}
\usage{
mc_fit_comp(formula, x)
}
\arguments{
\item{formula}{one sided model formula}

\item{x}{data frame}
}
\description{
Fit Composition Model for Total Moose
}
