% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\alias{wzi}
\alias{loo}
\alias{zeroinfl2}
\alias{summary.non_zeroinfl}
\alias{print.summary.non_zeroinfl}
\alias{nobs.zeroinfl}
\title{Internal Functions: Models}
\usage{
wzi(object, pass_data = FALSE, ...)

loo(object, ...)

zeroinfl2(
  formula,
  data,
  subset,
  na.action,
  weights,
  offset,
  dist = "ZIP",
  link = c("logit", "probit", "cloglog"),
  control = NULL,
  model = TRUE,
  y = TRUE,
  x = FALSE,
  solveH = TRUE,
  robust = FALSE,
  ...
)

\method{summary}{non_zeroinfl}(object, ...)

\method{print}{summary.non_zeroinfl}(x, digits = max(3, getOption("digits") - 3), ...)

\method{nobs}{zeroinfl}(object, ...)
}
\arguments{
\item{object}{A model as returned by `zeroinfl2()`.}

\item{pass_data}{Logical, to pass the data or not.}

\item{...}{Other parameters passes to underlying functions.}

\item{formula}{Model formula as in `pscl::zeroinfl()`.}

\item{data}{Moose data frame.}

\item{subset, na.action, weights, offset, model, y}{Arguments as
in `pscl::zeroinfl()`.}

\item{dist}{Count distribution, one of `"ZIP"`, `"ZINB"`, `"P"`, `"NB"`.}

\item{link}{Link function for the zero model.}

\item{control}{See `pscl::zeroinfl.control()`.}

\item{x}{Arguments for `pscl::zeroinfl()` or a
model as returned by `zeroinfl2()` for the methods.}

\item{solveH}{Logical, to use robust matrix inversion to get VCV.}

\item{robust}{Logical, use CL/PL robust regression approach.}

\item{digits}{Digits for print method.}
}
\description{
These functions power the total Moose estimation and prediction.
}
\details{
`zeroinfl2` is a customized version of the `pscl::zeroinfl()` function,
but this also fits the non-ZI counterparts in a way that simplifies
downstream analyses (i.e. PI calculations). Intended for internal use.

`wzi` applies a leave-one-out approach to temper influential observations.
The process finds weights that are related to leverage
(how much each observation contributes to the model likelihood).

The robust version (CL/PL = conditional likelihood / pseudo likelihood)
applies the conditioning described in Solymos et al. (2021)
<https://doi.org/10.1002/env.1149>.
}
\keyword{internal}
\keyword{models}
\keyword{regression}
