% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total.R
\name{total}
\alias{total}
\alias{switch_response}
\alias{mc_update_total}
\alias{mc_fit_total}
\alias{mc_models_total}
\alias{mc_predict_total}
\alias{mc_get_pred}
\alias{pred_density_moose_PI}
\alias{mc_plot_residuals}
\alias{mc_plot_predpi}
\alias{mc_plot_pidistr}
\alias{mc_plot_predfit}
\title{Total Moose Workflow}
\usage{
switch_response(type = "total")

mc_update_total(x, srv = NULL, ss = NULL)

mc_fit_total(
  x,
  vars = NULL,
  zi_vars = NULL,
  dist = "ZINB",
  weighted = FALSE,
  robust = FALSE,
  intercept = c("both", "count", "zero", "none"),
  ...
)

mc_models_total(ml, x, coefs = TRUE)

mc_predict_total(model_id, ml, x, do_boot = TRUE, do_avg = FALSE)

mc_get_pred(PI, ss = NULL)

pred_density_moose_PI(PI)

mc_plot_residuals(model_id, ml, x)

mc_plot_predpi(PI)

mc_plot_pidistr(PI, id = NULL, plot = TRUE, breaks = "Sturges")

mc_plot_predfit(i, PI, interactive = FALSE)
}
\arguments{
\item{type}{The type of the response, can be `"total"` or
`"cows"` for `switch_response`.}

\item{x}{A data frame with Moose data,
or a data frame from `mc_update_total()`.}

\item{srv}{Logical vector, rows of `x` that are surveyed,
falls back to global options when `NULL`.}

\item{ss}{Logical vector to subset `x`, default is to take no subset.}

\item{vars}{column names of `x` to be used as predictors for
the count model.}

\item{zi_vars}{optional, column names of `x` to be used as
predictors for the zero model.}

\item{dist}{Count distribution (`P`, `NB`, `ZIP`, `ZINB`).}

\item{weighted}{Logical, to use weighting to moderate
influential observations.}

\item{robust}{Logical, use robust regression approach.}

\item{intercept}{Which intercepts to keep. Dropped intercepts lead to
regression through the origin (at the linear predictor scale).}

\item{...}{Other arguments passed to `zeroinfl2()`.}

\item{ml}{Named list of models.}

\item{coefs}{logical, return coefficient table too.}

\item{model_id}{model ID or model IDs (can be multiple from `names(ml)`).}

\item{do_boot}{Logical, to do bootstrap or not.}

\item{do_avg}{Logical, to do model averaging or not.}

\item{PI}{PI object returned by `mc_predict_total()`}

\item{id}{Cell ID.}

\item{plot}{Logical, to plot or just give summary.}

\item{breaks}{Breaks argument passed to `graphics::hist()`.}

\item{i}{Column (variable) name or index.}

\item{interactive}{Logical, draw interactive plot.}
}
\description{
`switch_response` switches between total Moose vs. cows only.
This sets the column name for total Moose estimation.
`mc_update_total` Updates/prepares the Moose data set for downstream analyses (i.e. calculates some derived variables, sets a surveyed/unsurveyed indicator, and optionally takes a subset).
`mc_fit_total` fit total Moose abundance models.
`mc_models_total` prints out estimates from the models.
}
\examples{

mc_options(B=10)

x <- read.csv(
    system.file("extdata/MayoMMU_QuerriedData.csv",
        package="moosecounter"))

#switch_response("cows") # for cows only
switch_response("total")

x <- mc_update_total(x)

mc_plot_univariate("Subalp_Shrub_250buf", x, "ZINB")

vars <- c("ELC_Subalpine", "Fire1982_2012", "Fire8212_DEM815",
    "NALC_Needle", "NALC_Shrub", "Subalp_Shrub_250buf",
    "ELCSub_Fire8212DEM815", "SubShrub250_Fire8212DEM815")

mc_plot_multivariate(vars, x)

ML <- list()
ML[["Model 0"]] <- mc_fit_total(x, dist="ZINB", weighted=TRUE)
ML[["Model 1"]] <- mc_fit_total(x, vars[1:2], dist="ZINB", weighted=TRUE)
ML[["Model 2"]] <- mc_fit_total(x, vars[2:3], dist="ZIP", weighted=TRUE)
ML[["Model 3"]] <- mc_fit_total(x, vars[3:4], dist="ZINB", weighted=TRUE)

mc_models_total(ML, x)
mc_plot_residuals("Model 3", ML, x)

PI <- mc_predict_total(
    model_id=c("Model 1", "Model 3"),
    ml=ML,
    x=x,
    do_boot=TRUE, do_avg=TRUE)

mc_get_pred(PI)
pred_density_moose_PI(PI)
mc_plot_predpi(PI)
mc_plot_pidistr(PI)
mc_plot_pidistr(PI, id=2)

}
\keyword{models}
\keyword{regression}
